refactor(type): split action reaction data types
