+++ Source/Portfolio/Type/CActionDataTypes.cpp
@@ -1,0 +1,17 @@
+#include "Type/CActionDataTypes.h"
+
+#include "Action/CAction.h"
+
+bool FActionData::IsValidMinimal() const
+{
+	return ActionDataKey.IsValidMinimal()
+		&& IsValid(ActionExecutorKey.Get())
+		&& IsValid(Montage);
+}
+
+bool FActionExecutionContext::IsValidMinimal() const
+{
+	return ActionDataKey.IsValidMinimal()
+		&& ActionData.IsValidMinimal()
+		&& IsValid(ActionExecutor);
+}
